Creation of ServiceCall entity, repository, and service. Implementation of POST and GETALL tests in the ServiceCall controller

+++ src/main/java/com/api/customer/entities/ServiceCall.java
@@ -31,7 +31,7 @@ public class ServiceCall {
     private String title;
     private BigDecimal value;
     private Integer idTechnician;
-    private Integer idCustomer;
+    private Long idCustomer;
     @Transient
     private String nameTechnician;
     @Transient
